Movie inherited tables. Genres
+ Entities: MovieActor, MovieDirector
таблиця Actor перейменована в Person (там будуть записи для усіх імен, акторів і режисерів)

+ enum MovieGenre для переліку жанрів фільмів. Також написано GenreConverter для перетворення рядка в код жанру

додано файли конфігурації для нових entities, похідних від фільмів

+++ Core/Entities/Movie.cs
@@ -5,9 +5,9 @@ public class Movie
     public int Id { get; set; }
 
     public string Name { get; set; } = null!;
-    public short DurationMinutes { get; set; }
+    public byte DurationMinutes { get; set; }
     public byte AgeLimit { get; set; }
-    public byte Genre { get; set; }
+    public ulong Genre { get; set; }
 
     public string? Description { get; set; }
     public DateOnly ReleaseDate { get; set; }
